fix: resolve all compile errors and warnings
- Merge orphaned methods back into PayrollService impl block
- Change i128 to i64 for total_amount/amount (Postgres BIGINT compat)
- Fix use-after-move of payroll.id in execute_payroll
- Remove unused imports: Uuid, Serialize in multiple files
- Update zk_prover/stellar_service signatures to use i64 consistently

+++ src/api/employee.rs
@@ -4,7 +4,7 @@ use axum::{
     Json,
     Extension,
 };
-use serde::{Deserialize, Serialize};
+use serde::Deserialize;
 use serde_json::{json, Value};
 use std::sync::Arc;
 
@@ -16,7 +16,7 @@ pub struct AddEmployeeRequest {
     pub wallet_address: String,
     pub department: String,
     /// Salary in stroops (1 XLM = 10_000_000 stroops)
-    pub salary: i128,
+    pub salary: i64,
     pub randomness: Vec<u8>,
 }
